cloud_functions/add-weights/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage
google-cloud-firestore
google-cloud-pubsub